Use unminified version to avoid optimization errors
It appears that using the minified mermaidjs version may
result in some functions being optimized out dependending
on the concrete VuePress configuration. Using the unminified
version as import does not yield that result.

I have no idea how to address this so help would be welcome!
For the time being, albeit adding massive overload to the
site, this works for all the VuePress setups I can think of
at the moment.